reduce unsafe blocks by implementing TrustedLen trait
