+++ polars/polars-arrow/Cargo.toml
@@ -10,6 +10,7 @@ description = "Arrow interfaces for Polars DataFrame library"
 
 [dependencies]
 arrow = { package = "arrow2", git = "https://github.com/jorgecarleitao/arrow2", rev = "b617331354fd8c64c2126b6f4fc6f9935f7971ab", default-features = false }
+hashbrown = "0.11"
 # arrow = { package = "arrow2", git = "https://github.com/ritchie46/arrow2", default-features = false, features = ["compute"], branch = "offset_pub" }
 # arrow = { package = "arrow2", version = "0.8", default-features = false }
 num = "^0.4"
