+++ polars/polars-arrow/src/lib.rs
@@ -3,6 +3,7 @@ pub mod bit_util;
 #[cfg(feature = "compute")]
 pub mod compute;
 pub mod error;
+pub mod export;
 pub mod index;
 pub mod io;
 pub mod is_valid;
@@ -10,5 +11,3 @@ pub mod kernels;
 pub mod prelude;
 pub mod trusted_len;
 pub mod utils;
-
-pub use arrow;
