+++ polars/polars-arrow/src/trusted_len/mod.rs
@@ -1,9 +1,12 @@
 mod boolean;
+mod push_unchecked;
 mod rev;
 
 use crate::utils::TrustMyLength;
 use arrow::bitmap::utils::{BitmapIter, ZipValidity};
+pub use push_unchecked::*;
 pub use rev::FromIteratorReversed;
+use std::iter::Scan;
 use std::slice::Iter;
 
 /// An iterator of known, fixed size.
@@ -61,58 +64,11 @@ unsafe impl<T, I: TrustedLen + Iterator<Item = T>> TrustedLen for ZipValidity<'_
 unsafe impl TrustedLen for BitmapIter<'_> {}
 unsafe impl<A: TrustedLen> TrustedLen for std::iter::StepBy<A> {}
 
-pub trait PushUnchecked<T> {
-    /// Will push an item and not check if there is enough capacity
-    ///
-    /// # Safety
-    /// Caller must ensure the array has enough capacity to hold `T`.
-    unsafe fn push_unchecked(&mut self, value: T);
-
-    /// Will push an item and not check if there is enough capacity nor update the array's length
-    /// # Safety
-    /// Caller must ensure the array has enough capacity to hold `T`.
-    /// Caller must update the length when its done updating the vector.
-    unsafe fn push_unchecked_no_len_set(&mut self, value: T);
-
-    /// Extend the array with an iterator who's length can be trusted
-    fn extend_trusted_len<I: IntoIterator<Item = T> + TrustedLen>(&mut self, iter: I);
-
-    fn from_trusted_len_iter<I: IntoIterator<Item = T> + TrustedLen>(iter: I) -> Self;
+unsafe impl<I, St, F, B> TrustedLen for Scan<I, St, F>
+where
+    F: FnMut(&mut St, I::Item) -> Option<B>,
+    I: TrustedLen + Iterator<Item = B>,
+{
 }
 
-impl<T> PushUnchecked<T> for Vec<T> {
-    #[inline]
-    unsafe fn push_unchecked(&mut self, value: T) {
-        let end = self.as_mut_ptr().add(self.len());
-        std::ptr::write(end, value);
-        self.set_len(self.len() + 1);
-    }
-
-    #[inline]
-    unsafe fn push_unchecked_no_len_set(&mut self, value: T) {
-        let end = self.as_mut_ptr().add(self.len());
-        std::ptr::write(end, value);
-    }
-
-    #[inline]
-    fn extend_trusted_len<I: IntoIterator<Item = T> + TrustedLen>(&mut self, iter: I) {
-        let iter = iter.into_iter();
-        let upper = iter.size_hint().1.expect("must have an upper bound");
-        self.reserve(upper);
-
-        unsafe {
-            let mut dst = self.as_mut_ptr().add(self.len());
-            for value in iter {
-                std::ptr::write(dst, value);
-                dst = dst.add(1)
-            }
-            self.set_len(self.len() + upper)
-        }
-    }
-
-    fn from_trusted_len_iter<I: IntoIterator<Item = T> + TrustedLen>(iter: I) -> Self {
-        let mut v = vec![];
-        v.extend_trusted_len(iter);
-        v
-    }
-}
+unsafe impl<K, V> TrustedLen for hashbrown::hash_map::IntoIter<K, V> {}
