+++ polars/polars-core/Cargo.toml
@@ -135,7 +135,6 @@ ahash = "0.7"
 anyhow = "1.0"
 
 comfy-table = { version = "4.0", optional = true }
-hashbrown = { version = "0.11", features = ["rayon"] }
 jsonpath_lib = { version = "0.3.0", optional = true, git = "https://github.com/ritchie46/jsonpath", branch = "improve_compiled" }
 lazy_static = "1.4"
 ndarray = { version = "0.15", optional = true, default_features = false }
