+++ polars/polars-core/src/chunked_array/ops/is_in.rs
@@ -55,20 +55,16 @@ where
                 let mut ca: BooleanChunked = if self.len() == 1 && other.len() != 1 {
                     let value = self.get(0);
 
-                    // we know the iterators len
-                    unsafe {
-                        other
-                            .list()?
-                            .amortized_iter()
-                            .map(|opt_s| {
-                                opt_s.map(|s| {
-                                    let ca = s.as_ref().unpack::<T>().unwrap();
-                                    ca.into_iter().any(|a| a == value)
-                                }) == Some(true)
-                            })
-                            .trust_my_length(other.len())
-                            .collect_trusted()
-                    }
+                    other
+                        .list()?
+                        .amortized_iter()
+                        .map(|opt_s| {
+                            opt_s.map(|s| {
+                                let ca = s.as_ref().unpack::<T>().unwrap();
+                                ca.into_iter().any(|a| a == value)
+                            }) == Some(true)
+                        })
+                        .collect_trusted()
                 } else {
                     self.into_iter()
                         .zip(other.list()?.amortized_iter())
@@ -130,20 +126,16 @@ impl IsIn for Utf8Chunked {
             DataType::List(dt) if self.dtype() == &**dt => {
                 let mut ca: BooleanChunked = if self.len() == 1 && other.len() != 1 {
                     let value = self.get(0);
-                    // safety: we know the iterators len
-                    unsafe {
-                        other
-                            .list()?
-                            .amortized_iter()
-                            .map(|opt_s| {
-                                opt_s.map(|s| {
-                                    let ca = s.as_ref().unpack::<Utf8Type>().unwrap();
-                                    ca.into_iter().any(|a| a == value)
-                                }) == Some(true)
-                            })
-                            .trust_my_length(other.len())
-                            .collect_trusted()
-                    }
+                    other
+                        .list()?
+                        .amortized_iter()
+                        .map(|opt_s| {
+                            opt_s.map(|s| {
+                                let ca = s.as_ref().unpack::<Utf8Type>().unwrap();
+                                ca.into_iter().any(|a| a == value)
+                            }) == Some(true)
+                        })
+                        .collect_trusted()
                 } else {
                     self.into_iter()
                         .zip(other.list()?.amortized_iter())
