+++ polars/polars-core/src/frame/groupby/hashing.rs
@@ -38,16 +38,7 @@ where
         }
     });
 
-    let len = hash_tbl.len();
-
-    //  we know that iterators length
-    unsafe {
-        hash_tbl
-            .into_iter()
-            .map(|(_k, tpl)| tpl)
-            .trust_my_length(len)
-            .collect_trusted()
-    }
+    hash_tbl.into_iter().map(|(_k, tpl)| tpl).collect_trusted()
 }
 
 /// Determine group tuples over different threads. The hash of the key is used to determine the partitions.
@@ -108,15 +99,10 @@ where
                 });
                 offset += len;
             }
-            let len = hash_tbl.len();
-            //  we know that iterators length
-            unsafe {
-                hash_tbl
-                    .into_iter()
-                    .map(|(_k, v)| v)
-                    .trust_my_length(len)
-                    .collect_trusted::<Vec<_>>()
-            }
+            hash_tbl
+                .into_iter()
+                .map(|(_k, v)| v)
+                .collect_trusted::<Vec<_>>()
         })
     })
     .flatten()
